[CKSwiftXX] Adding test files for experimenting with Swift C++ Interop.
I am going to use the Module Map technique for doing interop here.

ComponentKit is an Xcode project but I do not know how to set things up
this way at the moment so for the time being I have a script that
creates a forest of symlinks to make the include paths work. The forest
consists of a self pointing symlink that has the name of whatever
component (so every directory with a header gets a directory symlink to itself
that has a name like "ComponentKit" or "RenderCore").

There is also a script that passes all the proper flags to compile a
swift file that imports various ComponentKit headers in C++ Interop
mode.